ARMV7: minor tweaks to inter-stage calling and payload handling.
Payloads, by design, can return. There's lots of mechanism in the payload code
to support it, and the chooser payload relies on it. Hence, we should not mark
the function call in exit_stage as noreturn.

Not all ARM have unified caches, and it's not always easy to tell what
to do. So we are very paranoid. Before we call between stages, we
should carefully flush the dcache to memory and invalidate the icache.
This may be more than is necessary on all architectures but it
doesn't really hurt for the most part.

So compile cache management code into all stages, and call the
flush dcache/invalidate icache from all stages.

+++ src/arch/armv7/lib/cache-cp15.c
@@ -46,8 +46,12 @@ static void set_section_dcache(int section, enum dcache_option option)
 	/*
 	 * FIXME(dhendrix): This calculation is from arch/arm/lib/board.c
 	 * in u-boot. We may need to subtract more due to logging.
+	 * FIXME(rminnich)
+	 * The cast avoids an incorrect overflow diagnostic.
+	 * We really need to start adding ULL to constants that are
+	 * intrinsically unsigned.
 	 */
-	tlb_addr = (CONFIG_SYS_SDRAM_BASE + (CONFIG_DRAM_SIZE_MB << 20UL));
+	tlb_addr = ((u32)CONFIG_SYS_SDRAM_BASE + (CONFIG_DRAM_SIZE_MB << 20UL));
 	tlb_addr -= tlb_size;
 	/* round down to next 64KB limit */
 	tlb_addr &= ~(0x10000 - 1);
